+++ tosshin_cpp/CMakeLists.txt
@@ -20,21 +20,6 @@ find_package(tosshin_interfaces REQUIRED)
 
 install(DIRECTORY "include" DESTINATION ".")
 
-add_library(${PROJECT_NAME}
-  "src/navigation_provider.cpp")
-
-target_include_directories(${PROJECT_NAME} PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-  $<INSTALL_INTERFACE:include>)
-
-ament_target_dependencies(${PROJECT_NAME} rclcpp tosshin_interfaces)
-
-install(TARGETS ${PROJECT_NAME}
-  EXPORT export_${PROJECT_NAME}
-  ARCHIVE DESTINATION "lib"
-  LIBRARY DESTINATION "lib"
-  RUNTIME DESTINATION "bin")
-
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
   ament_lint_auto_find_test_dependencies()
@@ -42,6 +27,5 @@ endif()
 
 ament_export_dependencies(rclcpp tosshin_interfaces)
 ament_export_include_directories("include")
-ament_export_libraries(${PROJECT_NAME})
 
 ament_package()
