+++ tosshin_cpp/include/tosshin_cpp/tosshin_cpp.hpp
@@ -23,4 +23,4 @@
 
 #include "./navigation_provider.hpp"
 
-#endif  // tosshin_CPP__TOSSHIN_CPP_HPP_
+#endif  // TOSSHIN_CPP__TOSSHIN_CPP_HPP_
