+++ tosshin_interfaces/package.xml
@@ -7,7 +7,7 @@
   <maintainer email="alfi.maulana.f@gmail.com">Alfi Maulana</maintainer>
   <license>MIT License</license>
   <buildtool_depend>ament_cmake</buildtool_depend>
-  <build_depend>rosidl_default_generator</build_depend>
+  <buildtool_depend>rosidl_default_generator</buildtool_depend>
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
   <exec_depend>rosidl_default_runtime</exec_depend>
